strip and chop up dependencies
