+++ submodules/ConversationFlows/util/rally/getColourForAttachmentResult.js
@@ -1,0 +1,5 @@
+const getColourForAttachmentResult = result => {
+  return result.DisplayColor ? result.DisplayColor : '#36a64f';
+};
+
+module.exports = getColourForAttachmentResult;
