+++ submodules/ConversationFlows/util/rally/shouldAddCommentForPrefix.js
@@ -1,0 +1,6 @@
+const shouldAddCommentForPrefix = IDprefix => {
+  if (IDprefix == 'TC') return false;
+  return true;
+};
+
+module.exports = shouldAddCommentForPrefix;
