+++ submodules/ConversationFlows/util/rally/shouldShowFooter.js
@@ -1,0 +1,14 @@
+const shouldShowFooter = idPrefix => {
+  switch (idPrefix) {
+    case 'TC':
+      return false;
+
+    case 'TS':
+      return false;
+
+    default:
+      return true;
+  }
+};
+
+module.exports = shouldShowFooter;
